itp: Create instr if not in cache
While we assumed at first that CH8 cannot stop on
an odd address, this appears to not be the case.
We construct a naive cache for all even pairs of instructions,
and if we look up a cache key
that does not exist, simply create an instruction
and store it in the cache.